change np.floats to float

+++ grl/utils/math.py
@@ -84,7 +84,7 @@ def arg_mellowmax(
     axis_last = np.moveaxis(x, axis, -1)
     mm = mellowmax(axis_last, beta=beta, axis=-1)
     batch_adv = axis_last - np.broadcast_to(np.expand_dims(mm, -1), axis_last.shape)
-    batch_beta = np.empty(mm.shape, dtype=np.float32)
+    batch_beta = np.empty(mm.shape, dtype=float)
 
     # Beta is computed as the root of this function
     def f(y, adv):
